+++ src/Core/Dfe.Complete.Domain/Enums/NoteTaskIdentifier.cs
@@ -51,5 +51,10 @@ public enum NoteTaskIdentifier
 
     [Description("confirm_this_transfer_has_authority_to_proceed")]
     [DisplayDescription("Confirm this transfer has authority to proceed")]
-    ConfirmTransferHasAuthorityToProceed = 12
+    ConfirmTransferHasAuthorityToProceed = 12,
+
+    [Description("church_supplemental_agreement")]
+    [DisplayDescription("Church supplemental agreement")]
+    ChurchSupplementalAgreement = 13,
+    
 }