+++ src/Frontend/Dfe.Complete/Extensions/TempDataExtensions.cs
@@ -11,5 +11,10 @@ namespace Dfe.Complete.Extensions
 			tempData["NotificationTitle"] = notificationTitle;
 			tempData["NotificationMessage"] = notificationMessage;
 		}
-	}
+
+        public static void SetTaskSuccessNotification(this ITempDataDictionary tempData)
+        {
+            tempData.SetNotification(NotificationType.Success, "Success", "Task updated successfully");
+        }
+    }
 }