+++ src/Tests/Dfe.Complete.Api.Tests.Integration/Controllers/ProjectsController/TaskDataControllerTests.cs
@@ -643,5 +643,94 @@ namespace Dfe.Complete.Api.Tests.Integration.Controllers.ProjectsController
             Assert.NotNull(existingProject);
             Assert.True(existingProject.AllConditionsMet);
         }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateChurchSupplementalAgreementTaskAsync_ShouldUpdate_ConversionTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateChurchSupplementalAgreementTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<ConversionTasksData>();
+            dbContext.ConversionTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Conversion;
+            command.NotApplicable = false;
+            command.Received = true;
+            command.Cleared = true;
+            command.Signed = true;
+            command.SignedByDiocese = true;
+            command.Saved = false;
+            command.SentOrSaved = true;
+            command.SignedBySecretaryState = true;
+
+            // Act
+            await tasksDataClient.UpdateChurchSupplementalAgreementTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.ConversionTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.False(existingTaskData.ChurchSupplementalAgreementNotApplicable);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementReceived); 
+            Assert.True(existingTaskData.ChurchSupplementalAgreementCleared);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSigned);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSignedDiocese);
+            Assert.False(existingTaskData.ChurchSupplementalAgreementSaved);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSent);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSignedSecretaryState);
+        }
+        [Theory]
+        [CustomAutoData(typeof(CustomWebApplicationDbContextFactoryCustomization))]
+        public async Task UpdateChurchSupplementalAgreementTaskAsync_ShouldUpdate_TransferTaskData(
+            CustomWebApplicationDbContextFactory<Program> factory,
+            ITasksDataClient tasksDataClient,
+            UpdateChurchSupplementalAgreementTaskCommand command,
+            IFixture fixture)
+        {
+            // Arrange
+            factory.TestClaims = [new Claim(ClaimTypes.Role, ApiRoles.ReadRole), new Claim(ClaimTypes.Role, ApiRoles.UpdateRole), new Claim(ClaimTypes.Role, ApiRoles.WriteRole)];
+
+            var dbContext = factory.GetDbContext<CompleteContext>();
+
+            var taskData = fixture.Create<TransferTasksData>();
+            dbContext.TransferTasksData.Add(taskData);
+
+            await dbContext.SaveChangesAsync();
+            command.TaskDataId = new TaskDataId { Value = taskData.Id.Value };
+            command.ProjectType = ProjectType.Transfer;
+            command.NotApplicable = false;
+            command.Received = true;
+            command.Cleared = true;
+            command.Signed = true;
+            command.SignedByDiocese = true;
+            command.Saved = false;
+            command.SentOrSaved = true;
+            command.SignedBySecretaryState = true;
+
+            // Act
+            await tasksDataClient.UpdateChurchSupplementalAgreementTaskAsync(command, default);
+
+            // Assert
+            dbContext.ChangeTracker.Clear();
+            var existingTaskData = await dbContext.TransferTasksData.SingleOrDefaultAsync(x => x.Id == taskData.Id);
+            Assert.NotNull(existingTaskData);
+            Assert.False(existingTaskData.ChurchSupplementalAgreementNotApplicable);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementReceived);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementCleared);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSignedIncomingTrust);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSignedDiocese);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSavedAfterSigningBySecretaryState);
+            Assert.False(existingTaskData.ChurchSupplementalAgreementSavedAfterSigningByTrustDiocese);
+            Assert.True(existingTaskData.ChurchSupplementalAgreementSignedSecretaryState);
+
+        }
     }
 }